docker-geoloc-main/app/requirements.txt
numpy == 1.20.2
scipy == 1.3.1
cython == 0.29.21
pandas == 1.0.5
sqlalchemy == 1.3.18
paho-mqtt == 1.5.0
psycopg2 == 2.8.5
six == 1.15.0
requests == 2.24.0
pyproj == 2.6.1
statistics == 1.0.3.5
matplotlib == 3.2.2
pykalman